Moving markdown function to a helpers file
